+++ src/components/transfer/stepper/TransferStepperNextButton.js
@@ -1,7 +1,7 @@
 import Button from "@material-ui/core/Button/Button";
 import React from "react";
 import makeStyles from "@material-ui/core/styles/makeStyles";
-import colors from "../../styles/colors";
+import colors from "../../../styles/colors";
 
 const useStyles = makeStyles(theme => ({
   button: {
