+++ src/components/transfer/stepper/TransferStepperSelectResource.js
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { jsx, css } from "@emotion/core";
 import { useSelector } from "react-redux";
-import text from '../../styles/text';
+import text from '../../../styles/text';
 
 export default function TransferStepperSelectResource() {
   const selectedTransferResource = useSelector(state => state.resources.selectedTransferResource);
