+++ src/components/transfer/stepper/TransferStepperToken.js
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import {jsx} from "@emotion/core";
 import { useState } from "react";
-import SearchTextField from "../widgets/text_fields/SearchTextField";
+import SearchTextField from "../../widgets/text_fields/SearchTextField";
 import { withStyles } from "@material-ui/styles";
 
 const TokenTextField = withStyles({
